+++ packages/gamut/src/GridForm/types.ts
@@ -1,4 +1,4 @@
-import { ValidationOptions } from 'react-hook-form';
+import { FormContextValues, ValidationOptions } from 'react-hook-form';
 
 import { ColumnSizes, ResponsiveProperty } from '../Layout';
 
@@ -11,7 +11,24 @@ export type GridFormCheckboxField = BaseFormField & {
   description: string;
   defaultValue?: boolean;
   label?: string;
+  validation?: Pick<ValidationOptions, 'required'>;
   type: 'checkbox';
+};
+
+export type GridFormCustomFieldProps = {
+  className: string;
+  error?: string;
+  field: GridFormCustomField;
+  register: FormContextValues['register'];
+  setValue: (value: any) => void;
+};
+
+export type GridFormCustomField = BaseFormField & {
+  defaultValue?: any;
+  label?: string;
+  render: (props: GridFormCustomFieldProps) => React.ReactNode;
+  validation?: ValidationOptions;
+  type: 'custom';
 };
 
 export type GridFormTextField = BaseFormField & {
@@ -24,11 +41,29 @@ export type GridFormTextField = BaseFormField & {
 export type GridFormSelectField = BaseFormField & {
   defaultValue?: string;
   label: string;
-  options: string[];
+  options: string[] | Record<string, number | string>;
+  validation?: Pick<ValidationOptions, 'required'>;
   type: 'select';
+};
+
+export type GridFormFileField = BaseFormField & {
+  defaultValue?: FileList;
+  label: string;
+  validation?: ValidationOptions;
+  type: 'file';
+};
+
+export type GridFormTextAreaField = BaseFormField & {
+  defaultValue?: string;
+  label: string;
+  validation?: ValidationOptions;
+  type: 'textarea';
 };
 
 export type GridFormField =
   | GridFormCheckboxField
+  | GridFormCustomField
   | GridFormTextField
-  | GridFormSelectField;
+  | GridFormSelectField
+  | GridFormFileField
+  | GridFormTextAreaField;