+++ packages/gamut/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './createPromise';
+export * from './omitProps';